Add pickle/unpickle support for structs
Structs are registered with a pickler/unpickler implementation as part
of a type registry. This adds two new opcodes that aren't supported in
mainline pickle (NEWSTRUCT & BUILDSTRUCT). We don't bump the protocol in
this case, since pickle will likely bump their protocol at some point
to. It's not clear what's best here, punting for now.