buffer_callback -> collect_buffers
When working with `PickleBuffer` objects, most of the time we want to
accumulate them in a list - `buffer_callback=buffers.append` is a common
pattern. To simplify the user interface and provide some mild speedups,
we change the signature accept a boolean `collect_buffers` kwarg. If
true, buffers are accumulated in a list and returned as a second return
value from `dumps`. If no buffers are found, `None` is returned instead.