Reuse memo table object for encoder type registry
The hashtable acting as a type registry:
- needs to be a map between python objects and positive `Py_ssize_t` values.
- is read only (after setup)
- doesn't need to support deletions
- only needs to support lookups by id)

This makes python dicts a bit overkill for our needs. Luckily the
existing `MemoTable` objects fit these needs perfectly.

This renames `MemoTable` to `LookupTable`, and updates the `Encoder` to
use `LookupTable` for the type registry.

This makes it slightly slower to create an `Encoder` (we do more work
upfront), with the benefit of serialization being slightly faster.
Another benefit is that since we iterate over the passed in `registry`
now, we can validate the arguments on init rather on calls to `dumps`,
providing slightly better error messages.