Mimimize GC usage for Struct types
This delays tracking Struct objects in Python's GC until a chance for
a reference cycle has been created. A similar optimization is already
used in cpython for dicts.

At the start of `Struct.__new__`, a struct type is tracked (like all GC
types). If at the end of `__new__` no GC objects have been added as
attributes to the struct, then the struct is untracked. Later, if a GC
tracked object is setattr'd on a struct, then the struct becomes tracked
again.

This reduces GC pressure for large amounts of Struct objects, and can
result in measurable speedups under load.