Access attributes via offset rather than name
Since we know all the offsets of the fields stored on a struct, we can
access the fields via offset rather than name in:
- `Struct.__new__`
- `Struct.__repr__`
- `Struct.__copy__`
- `Struct.__reduce__`
- `Struct.__eq__`/`Struct.__ne__`
- `Encoder.dumps` for structs
- `Decoder.loads` for structs

This is complicated, since the layout of `__slots__` members on a
PyObject is an internal detail, and isn't consistent with the order of
fields (slots are sorted per class). To handle this, we have to do a bit
of bookkeeping in the `StructMeta` to deduce and store the offset
information. It would be nice if we could avoid this, or implement our
own slots mechanism, but any variable-length type object would confuse
the existing slots mechanism, so that's a no-go.

This results in quite measurable speedups on real world workflows (up to
2x improvement). This is most noticeable when running with GC enabled,
which timeit disables by default. I've made GC-enabled the default for
the benchmarks now, and added a flag to disable it. This should provide
a better proxy for real world measurements, as most users run with the
Python GC enabled.